feat: show transfer info for offline subscriptions
