+++ src/features/user/Account/AccountHistory.module.scss
@@ -69,7 +69,6 @@
 
 .recurrency {
   width: 100%;
-  max-width: 600px;
   display: flex;
   flex-direction: column;
   justify-content: flex-start;
