+++ src/features/user/Account/components/AccountRecord.tsx
@@ -7,6 +7,7 @@ import { useTranslation } from 'next-i18next';
 import { TFunction } from 'next-i18next';
 import DownloadCodes from './DownloadCodes';
 import BackButton from '../../../../../public/assets/images/icons/BackButton';
+import TransferDetails from './TransferDetails';
 
 interface CommonProps {
   handleRecordToggle: (index: number | undefined) => void;
@@ -452,53 +453,6 @@ export function BankDetails({ recipientBank }: BankDetailsProps): ReactElement {
           <p>{formatDate(recipientBank.updated)}</p>
         </div>
       )}
-    </>
-  );
-}
-
-interface TransferDetailsProps {
-  account: Payments.BankAccount;
-}
-
-export function TransferDetails({
-  account,
-}: TransferDetailsProps): ReactElement {
-  const { t } = useTranslation(['me']);
-  return (
-    <>
-      <div className={styles.title}>{t('transferDetails')}</div>
-      <div className={styles.detailGrid}>
-        {account.beneficiary && (
-          <div className={styles.singleDetail}>
-            <p className={styles.title}>{t('beneficiary')}</p>
-            <p>{account.beneficiary}</p>
-          </div>
-        )}
-        {account.iban && (
-          <div className={styles.singleDetail}>
-            <p className={styles.title}>{t('iban')}</p>
-            <p>{account.iban}</p>
-          </div>
-        )}
-        {account.bic && (
-          <div className={styles.singleDetail}>
-            <p className={styles.title}>{t('bic')}</p>
-            <p>{account.bic}</p>
-          </div>
-        )}
-        {account.bankName && (
-          <div className={styles.singleDetail}>
-            <p className={styles.title}>{t('bankName')}</p>
-            <p>{account.bankName}</p>
-          </div>
-        )}
-        {account.swift && account?.swift !== 'swift' && (
-          <div className={styles.singleDetail}>
-            <p className={styles.title}>{t('swift')}</p>
-            <p>{account.swift}</p>
-          </div>
-        )}
-      </div>
     </>
   );
 }
